requirements.txt
greenlet==3.1.1
imbalanced-learn==0.12.4
joblib==1.4.2
numpy==1.24.3
nvidia-nccl-cu12==2.23.4
pandas==2.2.3
psycopg2-binary==2.9.9
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
SQLAlchemy==2.0.35
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
xgboost==2.1.1
